Move AssembleAppFQDN to lib/utils
